[!!!][BUGFIX] Move appliation to subfolder
Egroupware doesn't support beeing called directly on a domain (e.g.
https://egroupware.example.com/) or there's a bug causing this.
In an empty installation we experienced missing icons in the tree.
Therefore it's necessary to use a subfolder (e.g.
https://egroupware.example.com/egroupware).

This is a breaking change. You need to change the `webserver_url` in
header admin (https://egroupware.example.com/egroupware/setup`) to
`/egroupware`.